fix CI: drop useless static_cast<int64_t> in OpenBoard's ActionKeyTraits::key
GCC's -Wuseless-cast (enabled on the "Application ladder" CI leg's
gcc-debug build, -Werror) correctly flagged this: ProjectId::operator*()
already returns std::int64_t directly (KANBAN_DEFINE_STRONG_ID's
generated dereference operator), so casting its result to
std::int64_t again was a genuine no-op. keyToString<K> is a template
taking any key type by const-ref, so passing the already-int64_t value
directly is equivalent and compiles identically under every other
build's flags.

This is the cosmetic Minor the Task 9 SDD ledger entry noted and
deferred to the final review ("redundant static_cast in the key
extraction") -- deferring it correctly assumed it was cosmetic under
every warning set exercised so far; GCC's stricter set (not previously
exercised against this branch until this CI run) makes it a hard
Werror failure instead, so it needs fixing now rather than staying
deferred.

+++ examples/kanban/include/kanban/models/board_model.hpp
@@ -258,7 +258,7 @@ struct morph::model::ActionKeyTraits<kanban::OpenBoard> {
     static constexpr bool hasKey = true;
     static constexpr bool fromResult = false;
     static std::string key(const kanban::OpenBoard& action) {
-        return morph::model::keyToString(static_cast<std::int64_t>(*action.projectId));
+        return morph::model::keyToString(*action.projectId);
     }
 };
 template <>
